pyproject.toml
[tool.poetry]
name = "alimentaçãosocial"
version = "0.1.0"
description = ""
authors = ["ErnestoTSantos <ernesto.terra2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
black = "24.4.2"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
django = "5.0.6"
django-rest-framework = "0.1.0"
djangorestframework = "3.15.1"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.14.0"
flake8 = "7.0.0"
idna = "3.7"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
mccabe = "0.7.0"
more-itertools = "10.2.0"
msgpack = "1.0.8"
mypy-extensions = "1.0.0"
packaging = "24.0"
pathspec = "0.12.1"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "4.2.1"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.7.1"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pycodestyle = "2.11.1"
pyflakes = "3.2.0"
pyproject-hooks = "1.1.0"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.9.0"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sqlparse = "0.5.0"
tomlkit = "0.12.5"
trove-classifiers = "2024.4.10"
tzdata = "2024.1"
urllib3 = "2.2.1"
virtualenv = "20.26.1"
validate-email = "^1.3"
validate-docbr = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
